java program
Write a Java program that takes two strings as input. The program should check each character of the first string to see if it appears in the second string. If a character from the first string does not appear in the second string, it should be printed. If all characters of the first string appear in the second string, nothing is printed.

Input:
hello
world

Output:
he