java program
Given an array arr of n positive integers and an integer k, Your task is to return k largest elements in decreasing order. 
Input: arr[] = {12, 5, 787, 1, 23}, n = 5, k = 2
Output: 787 23
Explanation: 1st largest element in the array is 787 and second largest is 23.
Input: arr[] = {1, 23, 12, 9, 30, 2, 50}, n = 7, k = 3 
Output: 50 30 23
Explanation: 3 Largest element in the array are 50, 30 and 23.