java program
Given an array of integers where every element appears an even number of times except one element that appears an odd number of times, write a program to find that odd occurring element . The equal elements must appear in pairs in the array but there cannot be more than two consecutive occurrences of an element. 

5 
2 2 3 1 1

 It is valid and the odd occurring element present in it is 3. 

The output is 3.